Move nonlinear::Missing*::execute to .cc
